Make height of application match viewport, with scrollable panes
Height of the whole application is now set to match the viewport height.
All subwindows in the application handle this by having scrollable
areas inside of them. Sidebar now works like the note list already did
(a static header and a scrollable content area), and this is achieved
quite simply with the following pattern:

.layout                      <display flex, height 100vh>
  .container                 <display flex, flex-direction column>
    .static-header
    .scrollable-content      <overflow-y auto>

It gets trickier with the problem list and attachment list where we,
inside of the BasePane component's content area, want a static header
and footer and a scrollable list that takes up the remaining height.
Meaning, the scrollable area is now one level further down from the
previous example. In order for this to work we need the following
pattern to communicate the height down through the hierarchy:

.layout                      <display flex, height 100vh>
  .container                 <display flex, flex-direction column>
    .static-header
    .static-content          <display flex, flex-direction column, flex-grow 1>
      .wrapper               <display flex, flex-direction column, height 100%>
        .static-header
        .scrollable-content  <flex-grow 1, overflow-y auto>
        .static-footer

SIDE EFFECT OF THIS COMMIT:

Unfortunately, setting overflow-y auto on the problem list causes the
selected problem highlight effect, which is positioned a few pixels
outside of the list item, to get clipped out by the surroundings. This is
due to overflow-x automatically being set to auto when overflow-y is set
to auto. Apparently, this is according to the specification, so it is
not possible to have automatic overflow in one direction and visible in
the other. See https://css-tricks.com/popping-hidden-overflow/ for an
explanation. A workaround could be to involve javascript, but I chose to
remove the negative offset of the highlight effect instead.