Fix issue with auto height and scrollable panes from b41f6b6 not working
I thought that commit did what it said; "Make height of application match
viewport, with scrollable panes", but I must have had local changes that
I forgot to commit, because even when checking out that commit, the
problem and attachment lists do not become scrollable when the window
height decreases, as I had intended.

In my description in the commit message, it appears I missed the fact
that .static-content needs "overflow-y: auto" too. I'm not sure how my
local changes solved that, but the easiest way seems to be to simply use
the already existing "scrollable" attribute of the BasePane component,
which was added along with the note list. That attribute is now set to
true for all usages of BasePane, so perhaps it is actually redundant and
"overflow-y: auto" should be the default, but this will do for now.